add session 2

+++ app/controllers/admin.js
@@ -1,42 +1,53 @@
 const Work = require('../models/work');
 
 exports.getAddWork = (req, res, next) => {
-    res.render('add-work', {
-        pageTitle: 'Добавить новую работу в базу данных',
-        path: '/admin/add-work',
-    });
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        res.render('add-work', {
+            pageTitle: 'Добавить новую работу в базу данных',
+            path: '/admin/add-work',
+        });
+    }
 };
 
 
 exports.postAddWork = (req, res, next) => {
-    const work = new Work(
-        null,
-        req.body.title,
-        req.body.descript,
-        req.body.feature,
-        req.body.url,
-        req.body.year,
-        req.user._id
-    );
-
-    work.save();
-    res.redirect(`/admin/portfolio/${req.body.year}`);
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        const work = new Work(
+            null,
+            req.body.title,
+            req.body.descript,
+            req.body.feature,
+            req.body.url,
+            req.body.year,
+            req.user._id
+        );
+        work.save();
+        res.redirect(`/admin/portfolio/${req.body.year}`);
+    }
 };
 
 exports.listAddWork = (req, res, next) => {
-    Work.fetchAll()
-        .then(works => {
-            res.render('admin/portfolio', {
-                title: 'Admin - перечень выполненных работ',
-                path: 'admin/portfolio',
-                yearId: req.params.year,
-                prod: works,
-                hasProducts: works.length > 0
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        Work.fetchAll()
+            .then(works => {
+                res.render('admin/portfolio', {
+                    title: 'Admin - перечень выполненных работ',
+                    path: 'admin/portfolio',
+                    yearId: req.params.year,
+                    prod: works,
+                    hasProducts: works.length > 0
+                });
+            })
+            .catch(err => {
+                console.log(err);
             });
-        })
-        .catch(err => {
-            console.log(err);
-        });
+    }
 };
 
 exports.indexAddWork = (req, res, next) => {
@@ -48,37 +59,49 @@ exports.indexAddWork = (req, res, next) => {
 
 
 exports.editWork = (req, res, next) => {
-    Work.findById(req.params.id)
-        .then(work => {
-            res.render('admin/edit-work', {
-                work: work,
-                edit: true
-            });
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        Work.findById(req.params.id)
+            .then(work => {
+                res.render('admin/edit-work', {
+                    work: work,
+                    edit: true
+                });
 
-        });
+            });
+    }
 };
 
 exports.postEditWork = (req, res, next) => {
-    const work = new Work(
-        req.body.id,
-        req.body.title,
-        req.body.descript,
-        req.body.feature,
-        req.body.url,
-        req.body.year,
-        req.user._id
-    );
-    work.save()
-        .then(result => {
-            // console.log('rerod updated');
-            res.redirect(`/admin/portfolio/${req.body.year}`);
-        })
-        .catch(err => console.log(err));
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        const work = new Work(
+            req.body.id,
+            req.body.title,
+            req.body.descript,
+            req.body.feature,
+            req.body.url,
+            req.body.year,
+            req.user._id
+        );
+        work.save()
+            .then(result => {
+                // console.log('rerod updated');
+                res.redirect(`/admin/portfolio/${req.body.year}`);
+            })
+            .catch(err => console.log(err));
+    }
 
 };
 
 
 exports.deleteWork = (req, res, next) => {
-    Work.deleteById(req.body.workId);
-    res.redirect(`/admin/portfolio/${req.body.year}`);
+    if (!req.session.isLoggedIn) {
+        res.redirect('/admin');
+    } else {
+        Work.deleteById(req.body.workId);
+        res.redirect(`/admin/portfolio/${req.body.year}`);
+    }
 };
